+++ app/components/arrow-menu.tsx
@@ -2,13 +2,13 @@ import { Menu, paperClasses, styled } from "@mui/material";
 
 const commonBeforeStyles: Partial<CSSStyleDeclaration> = {
   content: '" "',
+  position: "absolute",
   display: "block",
-  background: "white",
   width: "10px",
   height: "10px",
-  transform: "rotate(45deg)",
-  position: "absolute",
   margin: "auto",
+  background: "white",
+  transform: "rotate(45deg)",
 };
 
 export const ArrowMenuTopCenter = styled(Menu)({
@@ -41,7 +41,7 @@ export const ArrowMenuTopBottom = styled(Menu)({
     "&:before": {
       ...commonBeforeStyles,
       top: -5,
-      left: "calc(100% - 40px)",
+      left: "0%",
       right: 0,
     },
   },
@@ -54,7 +54,7 @@ export const ArrowMenuBottomTop = styled(Menu)({
       ...commonBeforeStyles,
       bottom: -5,
       left: 0,
-      right: "calc(100% - 40px)",
+      right: "0%",
     },
   },
 });