+++ e2e/sorting.spec.ts
@@ -35,7 +35,7 @@ test("Segment sorting", async ({
     // Switch color on the first chart
     if (chartType === "Columns") {
       await within(selectors.edition.controlSectionByTitle("Segmentation"))
-        .getByText("None")
+        .getByLabelText("None")
         .click();
 
       await actions.mui.selectOption("Kanton");
@@ -50,7 +50,7 @@ test("Segment sorting", async ({
     expect(legendTexts[0]).toEqual("Zurich");
 
     await within(selectors.edition.controlSectionBySubtitle("Sort"))
-      .getByText("Automatic")
+      .getByLabelText("Automatic")
       .click();
 
     await actions.mui.selectOption("Name");
@@ -69,7 +69,7 @@ test("Segment sorting", async ({
     await screen.getByText("A → Z").click();
 
     await within(selectors.edition.controlSectionBySubtitle("Sort"))
-      .getByRole("button", { name: "Name" })
+      .getByLabelText("Name")
       .click();
 
     await actions.mui.selectOption("Automatic");
@@ -100,7 +100,7 @@ test("Segment sorting with hierarchy", async ({
   await sleep(3_000);
 
   const colorSection = selectors.edition.controlSectionByTitle("Segmentation");
-  await within(colorSection).getByText("None").click();
+  await within(colorSection).getByLabelText("None").click();
 
   await actions.mui.selectOption("Region");
   await selectors.chart.loaded();