+++ CHANGELOG.md
@@ -12,7 +12,8 @@ You can also check the
 # Unreleased
 
 - Styles
-  - Migrated header, footer, home and browse pages to the new CI / CD
+  - Migrated header, footer, home, browse and dataset preview pages to the new
+    CI / CD
   - Visualize font was changed to Noto Sans
 - Maintenance
   - Updated Next.js